Move fill area static function to Helper class
