+++ src/TSMapEditor/Helpers.cs
@@ -9,6 +9,7 @@ using System.IO;
 using TSMapEditor.GameMath;
 using TSMapEditor.Models;
 using TSMapEditor.Models.Enums;
+using TSMapEditor.Rendering;
 
 namespace TSMapEditor
 {
@@ -494,6 +495,84 @@ namespace TSMapEditor
             Point2D newcenter = new Point2D(firstNonTransparentX + width / 2, firstNonTransparentY + height / 2);
 
             return (texture, new Point2D(newcenter.X - oldcenter.X, newcenter.Y - oldcenter.Y));
+        }
+
+        /// <summary>
+        /// Gets map tile coordinates for an enclosed 'fill area' around a target tile.
+        /// </summary>
+        /// <param name="targetTile">Target map tile.</param>
+        /// <param name="map">Map instance.</param>
+        /// <param name="theaterGraphics">Theater graphics instance.</param>
+        /// <returns>Collection of map tile coordinates.</returns>
+        public static IEnumerable<Point2D> GetFillAreaTiles(MapTile targetTile, Map map, TheaterGraphics theaterGraphics)
+        {
+            TileImage tileGraphics = theaterGraphics.GetTileGraphics(targetTile.TileIndex);
+            MGTMPImage subCellImage = tileGraphics.TMPImages[targetTile.SubTileIndex];
+
+            byte terrainType = subCellImage.TmpImage.TerrainType;
+            int tileSetId = tileGraphics.TileSetId;
+            var tilesToCheck = new LinkedList<Point2D>(); // list of pending tiles to check
+            var tileCheckHashSet = new HashSet<int>();    // hash set of tiles that have been added to the list of
+                                                          // tiles to check at some point and so should not be added there again
+            tilesToCheck.AddFirst(targetTile.CoordsToPoint());
+            tileCheckHashSet.Add(targetTile.CoordsToPoint().GetHashCode());
+
+            var tilesToSkip = new HashSet<int>();         // tiles that have been confirmed as not being part of the area to fill
+
+            // tiles that have been confirmed as being part of the area to fill
+            var tilesToProcess = new List<Point2D>();
+
+            while (tilesToCheck.First != null)
+            {
+                var coords = tilesToCheck.First.Value;
+                tilesToCheck.RemoveFirst();
+
+                if (tilesToSkip.Contains(coords.GetHashCode()))
+                    continue;
+
+                var cell = map.GetTile(coords);
+                if (cell == null)
+                {
+                    tilesToSkip.Add(coords.GetHashCode());
+                    continue;
+                }
+
+                tileGraphics = theaterGraphics.GetTileGraphics(cell.TileIndex);
+                if (tileGraphics.TileSetId != tileSetId)
+                {
+                    tilesToSkip.Add(coords.GetHashCode());
+                    continue;
+                }
+
+                subCellImage = tileGraphics.TMPImages[cell.SubTileIndex];
+                if (subCellImage.TmpImage.TerrainType != terrainType)
+                {
+                    tilesToSkip.Add(coords.GetHashCode());
+                    continue;
+                }
+
+                // Mark this cell as one to process and nearby tiles as ones to check
+                tilesToProcess.Add(coords);
+
+                for (int y = -1; y <= 1; y++)
+                {
+                    for (int x = -1; x <= 1; x++)
+                    {
+                        if (y == 0 && x == 0)
+                            continue;
+
+                        var newCellCoords = new Point2D(x, y) + coords;
+                        int hash = newCellCoords.GetHashCode();
+                        if (!tilesToSkip.Contains(hash) && !tileCheckHashSet.Contains(hash))
+                        {
+                            tileCheckHashSet.Add(hash);
+                            tilesToCheck.AddLast(newCellCoords);
+                        }
+                    }
+                }
+            }
+
+            return tilesToProcess;
         }
     }
 }
